requirements.txt
obs-websocket-py==1.0
opencv-python==4.8.1.78
pyaudio==0.2.11
numpy==1.24.3
python-dotenv==1.0.0 